refactor(reset-password): merge all the methods into one file
