+++ .eslintrc.json
@@ -1,0 +1,52 @@
+{
+  "extends": "airbnb",
+  "env": {
+    "es6": true,
+    "browser": true,
+    "jest": true
+  },
+  "parser": "babel-eslint",
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "jsx": true
+    }
+  },
+  "plugins": [
+    "react"
+  ],
+  "rules": {
+    "semi": "off",
+    "no-nested-ternary": "off",
+    "max-len": "off",
+    "func-names": "off",
+    "no-restricted-syntax": "off",
+    "no-unused-expressions": "off",
+    "import/no-named-as-default": 0,
+    "no-empty": ["error", {
+      "allowEmptyCatch": true
+    }],
+    "import/no-extraneous-dependencies": [
+      "off",
+      {
+        "devDependencies": false,
+        "optionalDependencies": false,
+        "peerDependencies": false
+      }
+    ],
+    "react/forbid-prop-types": "off",
+
+    "react/jsx-filename-extension": [
+      1,
+      {
+        "extensions": [
+          ".js",
+          ".jsx"
+        ]
+      }
+    ],
+    "jsx-a11y/anchor-is-valid": "off",
+    "jsx-a11y/click-events-have-key-events": "off"
+  }
+}
